// src/shared/api/models/getApiGiftCardsListTypeParams.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiGiftCardsListTypeStatus } from './getApiGiftCardsListTypeStatus';
import type { GetApiGiftCardsListTypeSource } from './getApiGiftCardsListTypeSource';

export type GetApiGiftCardsListTypeParams = {
    /**
     * @minimum 1
     */
    page?: number;
    /**
     * @minimum 1
     * @maximum 100000
     */
    limit?: number;
    status?: GetApiGiftCardsListTypeStatus;
    /**
     * Filter by gift card source: ALL, MANUAL (created manually), ONLINE (purchased from online template)
     */
    source?: GetApiGiftCardsListTypeSource;
    /**
     * Search by gift card code, customer email, customer name, or recipient name
     */
    keyword?: string;
    fromDate?: string;
    toDate?: string;
};